refacto: better separation of concerns
