+++ api/src/api/mod.rs
@@ -1,20 +1,19 @@
+pub mod service;
+
 use actix_web::{post, web, Responder};
 use log::{debug, error, info, trace};
-use tokio::sync::Mutex;
 
 use crate::{
-    grpc_definitions::ExecuteResponse,
+    api::service::LambdoApiService,
     model::{RunRequest, RunResponse},
-    LambdoState,
+    vm_manager::grpc_definitions::ExecuteResponse,
 };
-use std::{error::Error, sync::Arc};
-
-use crate::service::run_code;
+use std::error::Error;
 
 #[post("/run")]
 async fn run(
     run_body: web::Json<RunRequest>,
-    state: web::Data<Arc<Mutex<LambdoState>>>,
+    state: web::Data<LambdoApiService>,
 ) -> Result<impl Responder, Box<dyn Error>> {
     debug!(
         "Received code execution request from http (language: {}, version: {})",
@@ -22,7 +21,7 @@ async fn run(
     );
     trace!("Request body: {:?}", run_body);
 
-    let response = run_code(state, run_body).await;
+    let response = state.run_code(run_body.into_inner()).await;
 
     let response = match response {
         Ok(response) => {
