+++ api/src/main.rs
@@ -1,11 +1,7 @@
+pub mod api;
 pub mod config;
-pub mod controller;
-pub mod grpc_definitions;
 pub mod model;
-pub mod net;
-pub mod service;
-pub mod state;
-pub mod vmm;
+pub mod vm_manager;
 
 use std::sync::Arc;
 
@@ -13,13 +9,14 @@ use config::LambdoConfig;
 use thiserror::Error;
 
 use crate::{
-    controller::run, grpc_definitions::lambdo_api_service_server::LambdoApiServiceServer,
-    state::LambdoState, vmm::vm_handler::VMHandler,
+    api::{run, service::LambdoApiService},
+    vm_manager::grpc_definitions::lambdo_api_service_server::LambdoApiServiceServer,
+    vm_manager::state::LambdoState,
+    vm_manager::VMListener,
 };
 use actix_web::{web, App, HttpServer};
 use clap::Parser;
 use log::{debug, error, info, trace};
-use net::setup_bridge;
 use tokio::sync::Mutex;
 
 #[derive(Parser)]
@@ -56,21 +53,29 @@ async fn main() -> std::io::Result<()> {
         config
     );
 
-    setup_bridge(&config.api.bridge, &config.api.bridge_address).unwrap();
-
+    info!("setting up");
     let lambdo_state = Arc::new(Mutex::new(LambdoState {
         vms: Vec::new(),
         config: config.clone(),
     }));
-
     let lambdo_state_clone = lambdo_state.clone();
+
+    let api_service = LambdoApiService::new_with_state(lambdo_state)
+        .await
+        .map_err(|e| {
+            error!("failed to set up API service: {}", e);
+            std::io::Error::new(std::io::ErrorKind::Other, e)
+        })?;
+
+    info!("everything is set up, starting servers");
+
+    let grpc_host = config.api.grpc_host.clone();
+    let grpc_port = config.api.gprc_port;
     // TODO: Shut down the web server when the gRPC server is down, and vice versa
     tokio::spawn(async move {
-        let grpc_host = config.api.grpc_host;
-        let grpc_port = config.api.gprc_port;
         let addr = format!("{}:{}", grpc_host, grpc_port).parse().unwrap();
         info!("Starting gRPC server on {}", addr);
-        let vm_handler = VMHandler::new(lambdo_state_clone);
+        let vm_handler = VMListener::new(lambdo_state_clone);
         tonic::transport::Server::builder()
             .add_service(LambdoApiServiceServer::new(vm_handler))
             .serve(addr)
@@ -81,12 +86,12 @@ async fn main() -> std::io::Result<()> {
             });
     });
 
-    let http_host = config.api.web_host;
+    let http_host = &config.api.web_host;
     let http_port = config.api.web_port;
+    let app_state = web::Data::new(api_service);
     info!("Starting web server on {}:{}", http_host, http_port);
-    let app_state = web::Data::new(lambdo_state);
     HttpServer::new(move || App::new().app_data(app_state.clone()).service(run))
-        .bind((http_host, http_port))?
+        .bind((http_host.clone(), http_port))?
         .run()
         .await
 }
