+++ api/src/model.rs
@@ -1,7 +1,8 @@
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::grpc_definitions::FileModel;
+use crate::config::LambdoLanguageConfig;
+use crate::vm_manager::grpc_definitions::FileModel;
 
 #[derive(Deserialize, Debug)]
 pub struct RunRequest {
@@ -16,4 +17,20 @@ pub struct RunResponse {
     pub status: u8,
     pub stdout: String,
     pub stderr: String,
+}
+
+pub struct LanguageSettings {
+    pub name: String,
+    pub version: String,
+    pub initramfs: String,
+}
+
+impl From<LambdoLanguageConfig> for LanguageSettings {
+    fn from(config: LambdoLanguageConfig) -> Self {
+        LanguageSettings {
+            name: config.name,
+            version: config.version,
+            initramfs: config.initramfs,
+        }
+    }
 }
