+++ api/src/vm_manager/vmm/grpc_definitions.rs
